test: BoardとgameCreateResのvalidateを追加
