+++ core/firebase.ts
@@ -9,6 +9,8 @@ import {
 
 import { reqEnv } from "./env.ts";
 
+import { sleep } from "../v1/test/client_util.ts";
+
 const firebaseConfig: FirebaseOptions = {
   apiKey: "AIzaSyBOas3O1fmIrl51n7I_hC09YCG0EEe7tlc",
   authDomain: "kakomimasu.firebaseapp.com",
@@ -20,27 +22,19 @@ const firebaseConfig: FirebaseOptions = {
   measurementId: "G-L43FT511YW",
 };
 
-type InitOption = {
-  auth?: boolean;
-  database?: boolean;
-};
-
-import { sleep } from "../v1/test/client_util.ts";
-export async function firebaseInit(
-  option: InitOption = { auth: true, database: true },
-) {
-  initializeApp(firebaseConfig);
+let app: ReturnType<typeof initializeApp>;
+export async function firebaseInit() {
+  if (app !== undefined) return;
+  app = initializeApp(firebaseConfig);
 
   if (reqEnv.FIREBASE_TEST) {
-    if (option.auth) {
-      const auth = getAuth();
-      connectAuthEmulator(auth, `http://${reqEnv.FIREBASE_EMULATOR_HOST}:9099`);
-      // connectAuthEmulatorの接続待ち(テスト時にleaking opsが発生するため)
-      await sleep(1000);
-    }
-    if (option.database) {
-      const db = getDatabase();
-      connectDatabaseEmulator(db, reqEnv.FIREBASE_EMULATOR_HOST, 9000);
-    }
+    const auth = getAuth();
+    connectAuthEmulator(auth, `http://${reqEnv.FIREBASE_EMULATOR_HOST}:9099`);
+
+    const db = getDatabase();
+    connectDatabaseEmulator(db, reqEnv.FIREBASE_EMULATOR_HOST, 9000);
+
+    // connectAuthEmulatorの接続待ち(テスト時にleaking opsが発生するため)
+    await sleep(1000);
   }
 }
