+++ deno.json
@@ -2,7 +2,8 @@
   "tasks": {
     "cache": "deno cache server.ts",
     "start": "deno run --allow-read --allow-env --allow-net server.ts",
-    "test": "deno test --allow-read --allow-env --allow-net"
+    "test": "deno test --allow-read --allow-env --allow-net",
+    "firebase-emu": "cd firebase-emulator && firebase emulators:start --import ./data/"
   },
   "fmt": {
     "files": { "exclude": ["firebase-emulator/"] }
